ISPN-14895 Distinct authorization tests by protocol
* Created listeners to capture the suite in use.
* Split the tests in different files.
