+++ server/testdriver/junit4/src/main/java/org/infinispan/server/test/junit4/InfinispanServerTestMethodRule.java
@@ -9,6 +9,7 @@ import org.infinispan.client.rest.configuration.RestClientConfigurationBuilder;
 import org.infinispan.counter.api.CounterManager;
 import org.infinispan.server.test.api.HotRodTestClientDriver;
 import org.infinispan.server.test.api.MemcachedTestClientDriver;
+import org.infinispan.server.test.api.RespTestClientDriver;
 import org.infinispan.server.test.api.RestTestClientDriver;
 import org.infinispan.server.test.api.TestClientDriver;
 import org.infinispan.server.test.core.TestClient;
@@ -47,6 +48,11 @@ public class InfinispanServerTestMethodRule implements TestRule, TestClientDrive
    @Override
    public RestTestClientDriver rest() {
       return testClient.rest();
+   }
+
+   @Override
+   public RespTestClientDriver resp() {
+      return testClient.resp();
    }
 
    @Override
