+++ server/tests/src/test/java/org/infinispan/server/security/authorization/AuthorizationLDAPIT.java
@@ -10,11 +10,15 @@ import org.infinispan.server.test.junit4.InfinispanServerTestMethodRule;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
 
 /**
  * @author Tristan Tarrant &lt;tristan@infinispan.org&gt;
  * @since 10.0
  **/
+@RunWith(AuthorizationSuiteRunner.class)
+@Suite.SuiteClasses({HotRodAuthorizationTest.class, RESTAuthorizationTest.class, RESPAuthorizationTest.class})
 @Category(Security.class)
 public class AuthorizationLDAPIT extends AbstractAuthorization {
    @ClassRule
